backend/requirements-optimize.txt
# Requirements for asset optimization script
Pillow>=9.0.0
